job-application-agent/main.py
from dotenv import laod_dotenv; load_dotenv()
from langchain_google_genai import ChatGoogleGenereativeAI
from langchain_core.prompts import ChatPromptTemplate
from langchain_core.messages import HumanMessage, AIMessage
from langchain_core.output_parsers import PydanticOutputParser
from langchain.agents import create_tool_calling_agent, AgentExecutor
from tools import tools
from schema import ApplicationOutput

