job-application-agent/requirements.txt
langchain
langchain-community
langchain-google-genai
pydantic
python-dotenv
